requirements.txt
apns2==0.5.0
asn1crypto==0.24.0
astroid==2.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.6.1
Django==2.2.1
django-cors-headers==3.0.1
django-graphos==0.3.41
django-qr-code==1.0.0
django-uuidfield==0.5.0
djangorestframework==3.9.3
djangorestframework-simplejwt==4.3.0
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
httpie==1.0.2
hyper==0.7.0
hyperframe==3.2.0
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
Pillow==6.0.0
py-vapid==1.5.0
pycparser==2.19
Pygments==2.4.0
PyJWT==1.7.1
pylint==2.2.2
pytz==2019.1
pywebpush==1.9.4
qrcode==6.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
wincertstore==0.2
wrapt==1.11.1
